+++ lib/Braintree/TransactionGateway.php
@@ -39,12 +39,13 @@ class TransactionGateway
     /**
      * @ignore
      * @access private
-     * @param array $attribs
+     * @param array $attribs (Note: $deviceSessionId and $fraudMerchantId params are deprecated. Use $deviceData instead)
      * @return Result\Successful|Result\Error
      */
     private function create($attribs)
     {
         Util::verifyKeys(self::createSignature(), $attribs);
+        $this->_checkForDeprecatedAttributes($attribs);
         return $this->_doCreate('/transactions', ['transaction' => $attribs]);
     }
 
@@ -78,8 +79,6 @@ class TransactionGateway
             'channel',
             'customerId',
             'deviceData',
-            'deviceSessionId',
-            'fraudMerchantId',
             'merchantAccountId',
             'orderId',
             'paymentMethodNonce',
@@ -101,9 +100,11 @@ class TransactionGateway
             'transactionSource',
             'type',
             'venmoSdkPaymentMethodCode',
+            'scaExemption',
             'shippingAmount',
             'discountAmount',
             'shipsFromPostalCode',
+            'deviceSessionId', 'fraudMerchantId', // NEXT_MAJOR_VERSION remove deviceSessionId and fraudMerchantId
             ['riskData' =>
                 [
                     //NEXT_MAJOR_VERSION remove snake case parameters, PHP should only accept camel case
@@ -259,7 +260,8 @@ class TransactionGateway
                 ['status' , 'previousNetworkTransactionId'],
             ],
             // NEXT_MAJOR_VERSION rename Android Pay to Google Pay
-            ['androidPayCard' => ['number', 'cryptogram', 'expirationMonth', 'expirationYear', 'eciIndicator', 'sourceCardType', 'sourceCardLastFour', 'googleTransactionId']]
+            ['androidPayCard' => ['number', 'cryptogram', 'expirationMonth', 'expirationYear', 'eciIndicator', 'sourceCardType', 'sourceCardLastFour', 'googleTransactionId']],
+            ['installments' => ['count']]
         ];
     }
 
@@ -571,6 +573,16 @@ class TransactionGateway
             throw new Exception\Unexpected(
             "Expected transaction or apiErrorResponse"
             );
+        }
+    }
+
+    private function _checkForDeprecatedAttributes($attributes)
+    {
+        if (isset($attributes['deviceSessionId'])) {
+            trigger_error('$deviceSessionId is deprecated, use $deviceData instead', E_USER_DEPRECATED);
+        }
+        if (isset($attributes['fraudMerchantId'])) {
+            trigger_error('$fraudMerchantId is deprecated, use $deviceData instead', E_USER_DEPRECATED);
         }
     }
 }